[#6] Add command history to REPL
REPLWidget now maintains a list of previous commands, along with an
iterator to the current command. When the user presses enter to send a
command, it is added near the front of the list, just after a default
empty string. When pressing up or down, the iterator is safely
incremented or decremented before filling the input box with the string.
The default empty string allows easy return to an empty input after
looking through the history.